+++ Relay/Relay/Login/RelayLoginViewController.swift
@@ -176,15 +176,3 @@ extension RelayLoginViewController : ASAuthorizationControllerDelegate {
         print("error \(error)")
     }
 }
-
-extension UIButton {
-    func setImage(systemName: String) {
-        contentHorizontalAlignment = .fill
-        contentVerticalAlignment = .fill
-        
-        imageView?.contentMode = .scaleAspectFit
-        
-        setImage(UIImage(systemName: systemName), for: .normal)
-    }
-}
-
